move newsletter type, add fetchEmail type
