+++ src/BannerNewsletter/index.tsx
@@ -3,7 +3,8 @@ import { css } from '@emotion/react';
 import { from, headline } from '@guardian/source-foundations';
 import { Button, SvgCross } from '@guardian/source-react-components';
 import { useEscapeShortcut, OnCloseClick, CLOSE_BUTTON_ID } from '../bannerCommon/bannerActions';
-import { NewsletterSubscribeCallback, CTA, NewsletterFrequency } from '../newsletterCommon';
+import { CTA, NewsletterFrequency } from '../newsletterCommon';
+import { NewsletterSubscribeCallback } from '../types/dcrTypes';
 import type { TrackClick } from '../utils/tracking';
 import { StyleData, selfServeStyles } from '../styles/bannerCommon';
 import { canRender, COMPONENT_NAME } from './canRender';
