+++ src/utils/StorybookWrapper.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { NewsletterSubscribeCallback } from '../NewsletterEpic';
+import { NewsletterSubscribeCallback } from '../types/dcrTypes';
 import { fontFaces, cssResets, previewStyles } from './styleUtilities';
 
 type Props = {
